pyproject.toml
[project]
name = "modelgrader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "Major Hayden", email = "major@mhtx.net" }]
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "ibm-watsonx-ai>=1.1.0",
    "google-generativeai>=0.8.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "rich>=13.9.0",
    "structlog>=24.4.0",
]

[project.scripts]
modelgrader = "modelgrader:main"

[build-system]
requires = ["uv_build>=0.8.23,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "fakeredis>=2.30.1",
    "pandas-stubs>=2.3.0.250703",
    "pre-commit>=4.2.0",
    "pyright>=1.1.403",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-cov>=6.2.1",
    "pytest-randomly>=3.16.0",
    "ruff>=0.12.5",
]

[tool.pytest.ini_options]
asyncio_default_fixture_loop_scope = "function"
addopts = [
    "--cov=modelgrader",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
    "--cov-branch",
    "--color=yes",
    "--verbose",
]
# Having problems with tests? Uncomment the following line to see more output.
# log_cli = false
# log_level = "DEBUG"
log_format = "%(asctime)s %(levelname)s %(message)s"
log_date_format = "%Y-%m-%d %H:%M:%S"
testpaths = ["tests"]

[tool.pyright]
venvPath = "."
venv = ".venv"

[tool.ruff]
fix = true

[tool.ruff.format]
preview = true
